Moved code and files so the game won't crash on servers anymore
