+++ .gitignore
@@ -1,24 +1,40 @@
-# Compiled class file
-*.class
+# gradle
 
-# Log file
-*.log
+.gradle/
+build/
+out/
+classes/
 
-# BlueJ files
-*.ctxt
+# eclipse
 
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
+*.launch
 
-# Package Files #
-*.jar
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
+# idea
 
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-replay_pid*
+.idea/
+*.iml
+*.ipr
+*.iws
+
+# vscode
+
+.settings/
+.vscode/
+bin/
+.classpath
+.project
+
+# macos
+
+*.DS_Store
+
+# fabric
+
+run/
+
+# java
+
+hs_err_*.log
+replay_*.log
+*.hprof
+*.jfr
