+++ src/main/java/nl/sniffiandros/bren/client/particle/MuzzleSmokeParticle.java
@@ -1,4 +1,4 @@
-package nl.sniffiandros.bren.common.particle;
+package nl.sniffiandros.bren.client.particle;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -6,13 +6,19 @@ import net.minecraft.client.particle.*;
 import net.minecraft.client.world.ClientWorld;
 import net.minecraft.particle.DefaultParticleType;
 
-public class AirRingParticle extends AscendingParticle {
+public class MuzzleSmokeParticle extends AscendingParticle {
 
-    protected AirRingParticle(ClientWorld world, double x, double y, double z, double velocityX, double velocityY, double velocityZ, float scaleMultiplier, SpriteProvider spriteProvider) {
-        super(world, x, y, z, 0.1F, 0.1F, 0.1F, velocityX, velocityY, velocityZ, scaleMultiplier, spriteProvider, 1.0F, 1, 0.0F, true);
-        this.setColor(2.0F,2.0F,2.0F);
+    protected MuzzleSmokeParticle(ClientWorld world, double x, double y, double z, double velocityX, double velocityY, double velocityZ, float scaleMultiplier, SpriteProvider spriteProvider) {
+        super(world, x, y, z, 0.1F, 0.1F, 0.1F, velocityX, velocityY, velocityZ, scaleMultiplier, spriteProvider, 1.0F, 2, 0.0F, true);
+        this.setColor(1.0F,1.0F,1.0F);
     }
 
+    @Override
+    public void tick() {
+        super.tick();
+        float f = (float) this.age/(this.getMaxAge()*2);
+        this.setColor(1.0F - f, 1.0F - f, 1.0F - f);
+    }
 
     public ParticleTextureSheet getType() {
         return ParticleTextureSheet.PARTICLE_SHEET_LIT;
@@ -31,7 +37,7 @@ public class AirRingParticle extends AscendingParticle {
         }
 
         public Particle createParticle(DefaultParticleType defaultParticleType, ClientWorld clientWorld, double d, double e, double f, double g, double h, double i) {
-            return new AirRingParticle(clientWorld, d, e, f, g, h, i, 1.8F, this.spriteProvider);
+            return new MuzzleSmokeParticle(clientWorld, d, e, f, g, h, i, 3.0F, this.spriteProvider);
         }
     }
 }
